restructured processing
